Adaptation for higher-clocked chips
Changed the TI_CC_SPIReadBurstReg function to fix several clock related
bugs. Previously, when using the code's logic on a faster chip like a
MSP430F5529, some registers where read wrongly since the code assumed
that some actions take long enough for the CPU to do some work in the
meantime.
Tested on the MSP430G2553 as well as the MSP430F5529 Launchpads with the
AIR Boosterpack.